assets_monitor/requirements.txt
amqp==5.2.0
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
frozendict==2.3.10
html5lib==1.1
idna==3.6
investpy==1.0.8
kombu==5.3.4
lxml==4.9.3
multitasking==0.0.11
numpy==1.26.2
pandas==2.1.3
peewee==3.17.0
prompt-toolkit==3.0.41
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
webencodings==0.5.1
yfinance==0.2.33
